requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.8
google-cloud-bigquery>=3.0
google-auth>=2.0
scikit-learn>=1.3
jupyter>=1.0
nbformat>=5.9
